+++ ql/src/java/org/apache/hadoop/hive/ql/exec/CopyTask.java
@@ -26,7 +26,6 @@ import org.apache.hadoop.fs.Path;
 import org.apache.hadoop.fs.PathFilter;
 import org.apache.hadoop.hive.common.FileUtils;
 import org.apache.hadoop.hive.ql.exec.repl.util.ReplUtils;
-import org.apache.hadoop.hive.ql.metadata.HiveException;
 import org.apache.hadoop.hive.ql.plan.CopyWork;
 import org.apache.hadoop.hive.ql.plan.api.StageType;
 import org.apache.hadoop.util.StringUtils;
@@ -44,7 +43,7 @@ public class CopyTask extends Task<CopyWork> implements Serializable {
   @Override
   public int execute() {
     try {
-      work.setValuesForDelayedExec();
+      work.setValuesBeforeExec();
     } catch (Exception e) {
       console.printError("Failed with exception " + e.getMessage(), "\n"
           + StringUtils.stringifyException(e));
