+++ ql/src/java/org/apache/hadoop/hive/ql/parse/ImportSemanticAnalyzer.java
@@ -37,6 +37,7 @@ import org.apache.hadoop.hive.metastore.api.Partition;
 import org.apache.hadoop.hive.metastore.api.StorageDescriptor;
 import org.apache.hadoop.hive.metastore.ReplChangeManager;
 import org.apache.hadoop.hive.metastore.utils.MetaStoreUtils;
+import org.apache.hadoop.hive.ql.Context;
 import org.apache.hadoop.hive.ql.ErrorMsg;
 import org.apache.hadoop.hive.metastore.txn.TxnUtils;
 import org.apache.hadoop.hive.ql.QueryState;
@@ -517,12 +518,13 @@ public class ImportSemanticAnalyzer extends BaseSemanticAnalyzer {
                                    Long writeId, int stmtId,
                                    String dumpRoot, ReplicationMetricCollector metricCollector) {
     Path dataPath = new Path(fromURI.toString(), EximUtil.DATA_PATH_NAME);
-    DelayExecUtil delayExecUtil = new DelayExecUtil(replace, writeId, stmtId, x.getHive(),
+    LoadTableStateWrapper loadTableStateWrapper = new LoadTableStateWrapper(replace, writeId, stmtId, x.getHive(),
         x.getCtx(), tblDesc, replicationSpec.isInReplicationScope());
 
     Task<?> copyTask;
-    // Corresponding work objects are not complete yet. Some of the values will be assigned when task is
-    // being executed.
+    // Corresponding work instances are not complete yet. Some of the values will be calculated and assigned when task
+    // is being executed. LoadTableStateWrapper contains the required variables (for this delayed calculation)
+    // and is passed in these work instances.
     if (replicationSpec.isInReplicationScope()) {
       boolean copyAtLoad = x.getConf().getBoolVar(HiveConf.ConfVars.REPL_RUN_DATA_COPY_TASKS_ON_TARGET);
       copyTask = ReplCopyTask.getLoadCopyTask(replicationSpec, dataPath, null, x.getConf(),
@@ -530,10 +532,10 @@ public class ImportSemanticAnalyzer extends BaseSemanticAnalyzer {
     } else {
       copyTask = TaskFactory.get(new CopyWork(dataPath, null, false, dumpRoot, metricCollector, true));
     }
-    ((CopyWork) copyTask.getWork()).setDelayExecUtil(delayExecUtil);
+    ((CopyWork) copyTask.getWork()).setLoadTableStateWrapper(loadTableStateWrapper);
 
     MoveWork moveWork = new MoveWork(x.getInputs(), x.getOutputs(), null, null, false,
-        dumpRoot, metricCollector, true, delayExecUtil);
+        dumpRoot, metricCollector, true, loadTableStateWrapper);
 
     //if Importing into existing table, FileFormat is checked by
     // ImportSemanticAnalzyer.checked checkTable()
@@ -1154,7 +1156,7 @@ public class ImportSemanticAnalyzer extends BaseSemanticAnalyzer {
     }
   }
 
-  static Table createNewTableMetadataObject(ImportTableDesc tblDesc, boolean isRepl)
+  private static Table createNewTableMetadataObject(ImportTableDesc tblDesc, boolean isRepl)
       throws SemanticException {
     Table newTable = new Table(tblDesc.getDatabaseName(), tblDesc.getTableName());
     //so that we know the type of table we are creating: acid/MM to match what was exported
@@ -1434,6 +1436,116 @@ public class ImportSemanticAnalyzer extends BaseSemanticAnalyzer {
       return db.getTable(tblDesc.getDatabaseName(),tblDesc.getTableName());
     } catch (InvalidTableException e) {
       return null;
+    }
+  }
+
+  public static class LoadTableStateWrapper {
+    private final boolean inReplScope;
+    private final boolean replace;
+    private final Long writeId;
+    private final int stmtId;
+    private final Hive hive;
+    private final Context ctx;
+    private final ImportTableDesc tblDesc;
+    private Path destPath = null, loadPath = null;
+    private LoadTableDesc.LoadFileType lft;
+    private boolean isSkipTrash = false;
+    private boolean needRecycle = false;
+    private final Path tgtPath;
+    private boolean isCalculated = false;
+
+    public LoadTableStateWrapper(boolean replace, Long writeId, int stmtId, Hive hive, Context ctx, ImportTableDesc tblDesc,
+        boolean inReplScope) {
+      this.replace = replace;
+      this.writeId = writeId;
+      this.stmtId = stmtId;
+      this.hive = hive;
+      this.ctx = ctx;
+      this.tblDesc = tblDesc;
+      this.inReplScope = inReplScope;
+      tgtPath = tblDesc == null ? null : new Path(tblDesc.getLocation());
+    }
+
+    public Table getTableIfExists() throws HiveException {
+      Table table = tableIfExists(tblDesc, hive);
+      if (table == null) {
+        table = createNewTableMetadataObject(tblDesc, true);
+      }
+
+      return table;
+    }
+
+    public void calculateValues(Table table) throws HiveException {
+      assert table != null;
+      assert table.getParameters() != null;
+
+      if (!isCalculated) {
+        if (inReplScope) {
+          isSkipTrash = MetaStoreUtils.isSkipTrash(table.getParameters());
+          if (table.isTemporary()) {
+            needRecycle = false;
+          } else {
+            org.apache.hadoop.hive.metastore.api.Database db = hive.getDatabase(table.getDbName());
+            needRecycle = db != null && ReplChangeManager.shouldEnableCm(db, table.getTTable());
+          }
+        }
+
+        if (AcidUtils.isTransactionalTable(table)) {
+          String mmSubdir = replace ? AcidUtils.baseDir(writeId) : AcidUtils.deltaSubdir(writeId, writeId, stmtId);
+          destPath = new Path(tgtPath, mmSubdir);
+          loadPath = tgtPath;
+          lft = LoadTableDesc.LoadFileType.KEEP_EXISTING;
+        } else {
+          destPath = loadPath = ctx.getExternalTmpPath(tgtPath);
+          lft = replace ? LoadTableDesc.LoadFileType.REPLACE_ALL : LoadTableDesc.LoadFileType.OVERWRITE_EXISTING;
+        }
+
+        isCalculated = true;
+      }
+    }
+
+    public Long getWriteId() {
+      return writeId;
+    }
+
+    public int getStmtId() {
+      return stmtId;
+    }
+
+    public boolean isReplace() {
+      return replace;
+    }
+
+    public Path getTgtPath() {
+      return tgtPath;
+    }
+
+    public boolean isInReplScope() {
+      return inReplScope;
+    }
+
+    public Path getDestPath() {
+      return destPath;
+    }
+
+    public Path getLoadPath() {
+      return loadPath;
+    }
+
+    public LoadTableDesc.LoadFileType getLft() {
+      return lft;
+    }
+
+    public boolean isSkipTrash() {
+      return isSkipTrash;
+    }
+
+    public boolean isNeedRecycle() {
+      return needRecycle;
+    }
+
+    public boolean isCalculated() {
+      return isCalculated;
     }
   }
